// src/shared/constants/constants.d.ts
export declare const BASE: number;
export declare const countReviewLines: number;
export declare const PROJECTS_DATA: {
    url: string;
    title: string;
};
export declare const CONTACTS_DATA: {
    url: string;
    title: string;
};
export declare const PRICES_DATA: {
    url: string;
    title: string;
};